fix(): accent for capitalized words
